Add `deleteTemporaryDirectory` to `PackageAnalyzer.inspectPackage`
Setting this to `false` retains the directory and prints its location to
the log.
Useful for debugging.

+++ lib/src/package_analyzer.dart
@@ -46,12 +46,12 @@ class PackageAnalyzer {
       new PackageAnalyzer(await DartSdk.create(sdkDir: sdkDir),
           flutterDir: flutterDir, pubCacheDir: pubCacheDir);
 
-  Future<Summary> inspectPackage(
-    String package, {
-    String version,
-    bool keepTransitiveLibs: false,
-    Logger logger,
-  }) async {
+  Future<Summary> inspectPackage(String package,
+      {String version,
+      bool keepTransitiveLibs: false,
+      Logger logger,
+      bool deleteTemporaryDirectory: true}) async {
+    deleteTemporaryDirectory ??= true;
     return withLogger(() async {
       log.info("Downloading package $package ${version ?? 'latest'}");
       String packageDir;
@@ -67,7 +67,12 @@ class PackageAnalyzer {
       try {
         return await _inspect(packageDir, keepTransitiveLibs);
       } finally {
-        await tempDir?.delete(recursive: true);
+        if (deleteTemporaryDirectory) {
+          await tempDir?.delete(recursive: true);
+        } else {
+          log.warning(
+              'Temporary directory was not deleted: `${tempDir.path}`.');
+        }
       }
     }, logger: logger);
   }
